backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
joblib
scikit-learn
xgboost